fix: Make StorePtr take a Register as value src
On AArch64, we cannot directly store an immediate to the stack, so
we must put it in a register first, and then using the reg to put it
on the stack.

This does however make the x86-64 implementation lack now, as it does
have `push imm` instruction (CISC), and makes the ir::Instruction now
ambigiously RISC and CISC.

Initially it was my idea to make the ir::Instruction encode as much logical
meaning as possible, so it was easy to understand. But having to enforce
the RISC rules (for ARM in this case) breaks that rule, and we now have to
depend on ISA-specific optimizations to optimize the chip completely.

So we must make ir::Instruction either:
- CISC, and lower it down at the ISA level if it is CISC
- RISC, and add another layer of optimizations per ISA even if those
  could've been consolidated with ir::Optimization